Use new BinderRecyclerAdapter for Channel items view.
